sytle(Validator): Validator 기능 별 분리
자동차 관련 검증과 시도 횟수 관련 검증 Validator를 분리했습니다.
